// src/types/product.ts
export interface Product {
  id: string;
  name: string;
  name_en: string;
  name_am: string;
  description: string;
  description_en: string;
  description_am: string;
  culturalStory_en?: string;
  culturalStory_am?: string;
  price: number;
  category: string;
  artisanId: string;
  artisanName: string;
  images: string[];
  isVerified: boolean;
  rating: number;
  status: 'Pending' | 'Approved' | 'Rejected';
  stock: number;
  sku: string;
  material?: string;
  material_en?: string;
  material_am?: string;
  careInstructions_en?: string;
  careInstructions_am?: string;
  isHandmade: boolean;
  productionTime: number;
  shippingLocations: string[];
  shippingCost: number;
  estimatedDelivery: string;
  returnPolicy: string;
  currency: string;
}